(1) move page component styling ts files under styles/pageComponentStyles directory
