+++ pages/soundWebtoons/[name]/[episode]/index.tsx
@@ -8,7 +8,7 @@ import {
   GridInfo,
   SoundLayer,
 } from "../../../../components/soundPart/soundLayer/soundLayer.component";
-import { PageContainer } from "./soundWebtoonIndex.styles";
+import { PageContainer } from "../../../../styles/pageComponentStyles/soundWebtoonIndex.styles";
 
 interface ITotalImageDimensionType {
   width: number;
